requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
numpy>=1.19.0
scikit-image>=0.17.0
plotly>=4.14.0
vit-pytorch>=0.19.0

# For tensorboard logging (optional, if you use --tensorboard)
tensorboardX>=2.1

# For image loading and saving
opencv-python>=4.5.0.52 